+++ lib/claper_web/router.ex
@@ -66,6 +66,7 @@ defmodule ClaperWeb.Router do
       live("/e/:code/manage/add/poll", EventLive.Manage, :add_poll)
       live("/e/:code/manage/edit/poll/:id", EventLive.Manage, :edit_poll)
       live("/e/:code/manage/add/form", EventLive.Manage, :add_form)
+      live("/e/:code/manage/import", EventLive.Manage, :import)
       live("/e/:code/manage/edit/form/:id", EventLive.Manage, :edit_form)
     end
   end