+++ CHANGELOG.md
@@ -37,3 +37,4 @@ _Security updates_
 - Migrate to Phoenix 1.7
 - Migrate to Liveview 0.18
 - Add multiple choice poll
+- Add feature to import all interactions from another presentation